refactor: restructure `tcads-client`
