--- packages/tcads-client/src/devices/mod.rs
@@ -1,2 +1,0 @@
-pub mod blocking;
-pub mod tokio;
