+++ packages/tcads-client/src/lib.rs
@@ -1,5 +1,6 @@
-pub mod devices;
+pub mod blocking;
 pub mod error;
+pub mod tokio;
 
 pub use tcads_core::{
     ads::{AdsReturnCode, IndexGroup, IndexOffset},
